various improvments, some pwm bug fixes

+++ utils.h
@@ -36,10 +36,10 @@ typedef struct
 
 int show_socket_error_reason(int socket);
 int initSPI(void);
-int mimas_start_stream(uint8_t start_bm, uint8_t proto_bm);
+int mimas_start_stream(uint16_t start_bm, uint16_t proto_bm);
 int mimas_store_packet(int chan, uint8_t* data, int len);
 int mimas_send_packet(int chan, uint8_t* data, int len);
-int mimas_refresh_start_stream(uint8_t start_bm, uint16_t proto_bm);
+int mimas_refresh_start_stream(uint16_t start_bm, uint16_t proto_bm);
 
 int mimas_store_pwm_val(uint8_t grp, int chan, uint16_t* val, uint8_t cnt);
 int mimas_store_pwm_period(uint8_t grp, uint16_t val);
